backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
python-dotenv==1.0.1
python-multipart==0.0.9
pdfplumber==0.11.4
sentence-transformers==2.2.2
numpy==1.24.3
supabase==2.3.0
openai==1.3.0
pydantic==2.5.0


